second ishauer clone

+++ userinfojpa/src/main/java/com/nr/isha/HomeController.java
@@ -6,6 +6,7 @@ import java.util.List;
 import org.apache.log4j.spi.LoggerFactory;
 import org.slf4j.Logger;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -13,6 +14,7 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.util.UriComponentsBuilder;
 
 import com.nr.isha.hibernate.repository.UserInterface;
 import com.nr.isha.service.UserService;
@@ -28,8 +30,16 @@ public class HomeController {
 	UserInterface userInterface;*/
 	 
 	 @RequestMapping(value = "/hello1")  
-	 	 public List<User> getTopic(){
-		return userService.getAlltopic();
+	 	 public ResponseEntity<List<User>> getTopic(){
+		 
+		  List<User> listofuser=userService.getAlltopic();
+		  if(listofuser==null){
+			  return new ResponseEntity("Unable to get data from. User  not found.",
+						HttpStatus.NOT_FOUND);
+			}else{
+				return new ResponseEntity<List<User>>(listofuser,HttpStatus.OK);
+		  }
+		  
 	 }
 	 
 	/* 
@@ -38,11 +48,23 @@ public class HomeController {
 		return  userService.gettopic(id);
 	 }*/
 	 
-	 @RequestMapping(method=RequestMethod.POST, value="/hello1")
-	 public void addUser(@RequestBody User user){
-		 System.out.println("HomeController->"+user);
-		 userService.add(user);
+	 @RequestMapping(method=RequestMethod.POST, value="/hello1/")
+	 public ResponseEntity<?>  addUser(@RequestBody User user, UriComponentsBuilder ucBuilder){
+		 System.out.println("HomeController->addUse->"+user);
+		 boolean verify=userService.isUserExist(user);
+		 if(verify){
+			 return new ResponseEntity("Unable to create. A User with name " + 
+						user.getName() + " already exist.",HttpStatus.CONFLICT);
+		 }else{
+			 userService.add(user);
+		 }
+		 HttpHeaders headers = new HttpHeaders();
+			headers.setLocation(ucBuilder.path("/hello1/{id}").buildAndExpand(user.getId()).toUri());
+			return new ResponseEntity<String>(headers, HttpStatus.CREATED);
+		 
+		  
 	 }
+	 
 	 
 	 @RequestMapping(method=RequestMethod.GET, value="/hello1/{name}/{password}")
 	 public ResponseEntity<?> isUserExist(@PathVariable("name")String name,@PathVariable("password")String password){
@@ -78,19 +100,40 @@ public class HomeController {
 	 */
 	 
 	 
-	/* @RequestMapping(method=RequestMethod.PUT, value="/hello1/{id}")
-	 public void update(@RequestBody User user,@PathVariable("id")Long id){
+	  @RequestMapping(method=RequestMethod.PUT, value="/hello1/{id}")
+	 public ResponseEntity<?> update(@RequestBody User user,@PathVariable("id")Long id){
 		 System.out.println(user);
-		 User user1=userService.gettopic(id);
 		 User usertemp=new User();
-		 usertemp.setId(user1.id);
-		 usertemp.setName(user1.name);
-		 usertemp.setPassword(user1.password);
-		 userService.update(usertemp);
-	 }*/
+		 User currentuser=userService.getUserById(id);
+		 if (currentuser == null) {
+				//logger.error("Unable to update. User with id {} not found.", id);
+				return new ResponseEntity("Unable to upate. User with id " + id + " not found.",
+						HttpStatus.NOT_FOUND);
+			}else{
+		 
+		 
+		System.out.println();
+		
+		 userService.update(user);
+			}
+		 return new ResponseEntity<User>(usertemp, HttpStatus.OK);
+	 } 
 	 
 	 @RequestMapping(method=RequestMethod.DELETE,value="/hello1/{id}")
-	 public void delete(@PathVariable("id")Long id){
-		  userService.delete(id);
+	 public ResponseEntity<?>  delete(@PathVariable("id")Long id){
+		 
+	 
+		  
+		  User user = userService.getUserById(id);
+			if (user == null) {
+			//	logger.error("Unable to delete. User with id {} not found.", id);
+				return new ResponseEntity("Unable to delete. User with id " + id + " not found.",
+						HttpStatus.NOT_FOUND);
+			}
+			if(user!=null){
+			userService.delete(id);
+			}
+			return new ResponseEntity<User>(HttpStatus.NO_CONTENT);
+		  
 	 }
 }
